[fix] 리디렉션 권한 버그 수정

+++ group/views.py
@@ -89,7 +89,10 @@ def info_nonadmin(request, group_id):
     user_state = MemberState.objects.filter(user=request.user,
                                             group_id=group_id).first()
 
-    if state == State.ADMIN:  # 운영진인 경우
+    if state == State.WITH_HISTORY:  # 이전 인증 내역이 있는 참여자
+        return redirect(f"/group/{group_id}/")
+
+    elif state == State.ADMIN:  # 운영진인 경우
         return redirect(f"/group/{group_id}/admin/")
 
     if request.method == "POST":
@@ -266,8 +269,9 @@ def preresult(request, group_id):
     members = MemberState.objects.filter(group=group)
     state = redirect_by_auth(request.user, group_id)
 
-    ctx = {"idea_list": idea_list, "members": members, "group": group}
-    return render(request, "preresult/preresult_admin.html", context=ctx)
+    if state == State.ADMIN:
+        ctx = {"idea_list": idea_list, "members": members, "group": group}
+        return render(request, "preresult/preresult_admin.html", context=ctx)
 
 
 def admin_page(request, group_id):
@@ -280,12 +284,16 @@ def admin_page(request, group_id):
     admin_set = set(admin_states_users)
     member_set = set(raw_member_states_users)
     member_states_users = list(member_set - admin_set)
-    ctx = {
-        "group_instance": group_instance,
-        "admin_states_users": admin_states_users,
-        "member_states_users": member_states_users,
-    }
-    return render(request, "admin/group_admin.html", ctx)
+
+    state = redirect_by_auth(request.user, group_id)
+
+    if state == State.ADMIN:
+        ctx = {
+            "group_instance": group_instance,
+            "admin_states_users": admin_states_users,
+            "member_states_users": member_states_users,
+        }
+        return render(request, "admin/group_admin.html", ctx)
 
 
 # 운영진&비운영진 멤버 리스트뽑는 함수
@@ -376,41 +384,43 @@ def preresult_modify(request, group_id):
     idea_list = Idea.objects.filter(
         group=group).order_by("-score")[:group.team_number]
     members = MemberState.objects.filter(group=group)
+    state = redirect_by_auth(request.user, group_id)
 
-    if request.method == "POST":
-        selected_values = request.POST.get("team_modify").split(",")
-        member_id = int(selected_values[0])
-        idea_id = int(selected_values[1])
-        mod_mem = MemberState.objects.get(id=member_id)
-        prev_idea = mod_mem.my_team_idea
-        mod_idea = Idea.objects.get(id=idea_id)
+    if state == State.ADMIN:
+        if request.method == "POST":
+            selected_values = request.POST.get("team_modify").split(",")
+            member_id = int(selected_values[0])
+            idea_id = int(selected_values[1])
+            mod_mem = MemberState.objects.get(id=member_id)
+            prev_idea = mod_mem.my_team_idea
+            mod_idea = Idea.objects.get(id=idea_id)
 
-        # (전제)팀장은 수정 페이지에서 팀 변경되면 안됨. 함수 실행X 바로 리디렉션
-        for idea in idea_list:
-            if idea.author == mod_mem.user:
-                url = reverse("group:preresult", args=[group.id])
-                return redirect(url)
+            # (전제)팀장은 수정 페이지에서 팀 변경되면 안됨. 함수 실행X 바로 리디렉션
+            for idea in idea_list:
+                if idea.author == mod_mem.user:
+                    url = reverse("group:preresult", args=[group.id])
+                    return redirect(url)
 
-        # (수정과정1)전에 있던 아이디어의 멤버에서 해당 멤버스테이트의 유저를 지움(수정한 아이디어의 멤버를 추가하기 위해서)
-        if prev_idea != mod_idea:
-            prev_idea.member.remove(mod_mem.user)
-            prev_idea.save()
-        # (수정과정2)해당 멤버스테이트의 최종 팀을 수정한 아이디어의 팀으로 설정
-        mod_mem.my_team_idea = mod_idea
-        mod_mem.save()
-        # (수정과정3)수정한 아이디어의 멤버에 해당 멤버스테이트의 유저를 추가.
-        mod_idea.member.add(mod_mem.user)
-        mod_idea.save()
+            # (수정과정1)전에 있던 아이디어의 멤버에서 해당 멤버스테이트의 유저를 지움(수정한 아이디어의 멤버를 추가하기 위해서)
+            if prev_idea != mod_idea:
+                prev_idea.member.remove(mod_mem.user)
+                prev_idea.save()
+            # (수정과정2)해당 멤버스테이트의 최종 팀을 수정한 아이디어의 팀으로 설정
+            mod_mem.my_team_idea = mod_idea
+            mod_mem.save()
+            # (수정과정3)수정한 아이디어의 멤버에 해당 멤버스테이트의 유저를 추가.
+            mod_idea.member.add(mod_mem.user)
+            mod_idea.save()
 
-        url = reverse("group:preresult", args=[group.id])
-        return redirect(url)
-    else:
-        ctx = {
-            "members": members,
-            "idea_list": idea_list,
-            "group": group,
-        }
-        return render(request, "preresult/preresult_modify.html", context=ctx)
+            url = reverse("group:preresult", args=[group.id])
+            return redirect(url)
+        else:
+            ctx = {
+                "members": members,
+                "idea_list": idea_list,
+                "group": group,
+            }
+            return render(request, "preresult/preresult_modify.html", context=ctx)
 
 
 def group_detail(request, group_id):
@@ -419,48 +429,52 @@ def group_detail(request, group_id):
     other_ideas = Idea.objects.filter(group=group).exclude(author=request.user)
     user_state = MemberState.objects.filter(user=request.user,
                                             group=group).first()
+    state = redirect_by_auth(request.user, group_id)
 
-    ideas_votes = {}
-    if user_state:
-        ideas_votes["idea_vote1_id"] = user_state.idea_vote1_id
-        ideas_votes["idea_vote2_id"] = user_state.idea_vote2_id
-        ideas_votes["idea_vote3_id"] = user_state.idea_vote3_id
+    if (state == State.WITH_HISTORY or state == State.ADMIN):
+        ideas_votes = {}
+        if user_state:
+            ideas_votes["idea_vote1_id"] = user_state.idea_vote1_id
+            ideas_votes["idea_vote2_id"] = user_state.idea_vote2_id
+            ideas_votes["idea_vote3_id"] = user_state.idea_vote3_id
 
-    has_voted = user_state and (user_state.idea_vote1 or user_state.idea_vote2
-                                or user_state.idea_vote3)
+        has_voted = user_state and (user_state.idea_vote1 or user_state.idea_vote2
+                                    or user_state.idea_vote3)
 
-    ctx = {
-        "group": group,
-        "author_ideas": author_ideas,
-        "other_ideas": other_ideas,
-        "ideas_votes": ideas_votes,
-        "has_voted": has_voted,
-    }
-    return render(request, "group/group_detail.html", ctx)
+        ctx = {
+            "group": group,
+            "author_ideas": author_ideas,
+            "other_ideas": other_ideas,
+            "ideas_votes": ideas_votes,
+            "has_voted": has_voted,
+        }
+        return render(request, "group/group_detail.html", ctx)
 
 
 def idea_create(request, group_id):
     group = get_object_or_404(Group, id=group_id)
+    state = redirect_by_auth(request.user, group_id)
 
-    if Idea.objects.filter(group=group, author=request.user).exists():
-        messages.error(request, "이미 이 그룹에 대한 아이디어를 제출했습니다.")
-        return redirect("group:group_detail", group_id=group.id)
-
-    if request.method == "POST":
-        form = IdeaForm(request.POST, request.FILES)
-        if form.is_valid():
-            idea = form.save(commit=False)
-            idea.group = group
-            idea.author = request.user
-            idea.save()
+    if state == State.WITH_HISTORY:
+        if Idea.objects.filter(group=group, author=request.user).exists():
+            messages.error(request, "이미 이 그룹에 대한 아이디어를 제출했습니다.")
             return redirect("group:group_detail", group_id=group.id)
-    else:
-        form = IdeaForm()
-    ctx = {
-        "form": form,
-        "group": group,
-    }
-    return render(request, "group/group_idea_create.html", ctx)
+
+        if request.method == "POST":
+            form = IdeaForm(request.POST, request.FILES)
+            if form.is_valid():
+                idea = form.save(commit=False)
+                idea.group = group
+                idea.author = request.user
+                idea.save()
+                return redirect("group:group_detail", group_id=group.id)
+        else:
+            form = IdeaForm()
+        ctx = {
+            "form": form,
+            "group": group,
+        }
+        return render(request, "group/group_idea_create.html", ctx)
 
 
 def idea_modify(request, group_id, idea_id):
@@ -469,23 +483,25 @@ def idea_modify(request, group_id, idea_id):
                              id=idea_id,
                              group=group,
                              author=request.user)
+    state = redirect_by_auth(request.user, group_id)
 
-    if request.method == "POST":
-        form = IdeaForm(request.POST, request.FILES, instance=idea)
-        if form.is_valid():
-            form.save()
-            return redirect("group:idea_detail",
-                            group_id=group.id,
-                            idea_id=idea.id)
-    else:
-        form = IdeaForm(instance=idea)
+    if state == State.WITH_HISTORY:
+        if request.method == "POST":
+            form = IdeaForm(request.POST, request.FILES, instance=idea)
+            if form.is_valid():
+                form.save()
+                return redirect("group:idea_detail",
+                                group_id=group.id,
+                                idea_id=idea.id)
+        else:
+            form = IdeaForm(instance=idea)
 
-    ctx = {
-        "form": form,
-        "group": group,
-        "idea": idea,
-    }
-    return render(request, "group/group_idea_modify.html", ctx)
+        ctx = {
+            "form": form,
+            "group": group,
+            "idea": idea,
+        }
+        return render(request, "group/group_idea_modify.html", ctx)
 
 
 def idea_delete(request, group_id, idea_id):
@@ -494,21 +510,25 @@ def idea_delete(request, group_id, idea_id):
                              id=idea_id,
                              group=group,
                              author=request.user)
+    state = redirect_by_auth(request.user, group_id)
 
-    if request.method == "POST" and request.POST.get("action") == "delete":
-        idea.delete()
-        return redirect("group:group_detail", group_id=group.id)
+    if state == State.WITH_HISTORY:
+        if request.method == "POST" and request.POST.get("action") == "delete":
+            idea.delete()
+            return redirect("group:group_detail", group_id=group.id)
 
 
 def idea_detail(request, group_id, idea_id):
     group = get_object_or_404(Group, id=group_id)
     idea = get_object_or_404(Idea, id=idea_id, group=group)
+    state = redirect_by_auth(request.user, group_id)
 
-    context = {
-        "group": group,
-        "idea": idea,
-    }
-    return render(request, "group/group_idea_detail.html", context)
+    if (state == State.WITH_HISTORY or state == State.ADMIN):
+        context = {
+            "group": group,
+            "idea": idea,
+        }
+        return render(request, "group/group_idea_detail.html", context)
 
 
 def idea_download(request, group_id, idea_id):
@@ -530,69 +550,71 @@ def idea_download(request, group_id, idea_id):
 def vote_create(request, group_id):
     group = Group.objects.get(pk=group_id)
     user = request.user
+    state = redirect_by_auth(user, group_id)
 
-    try:
-        user_state, created = MemberState.objects.get_or_create(user=user,
-                                                                group=group)
+    if state == State.WITH_HISTORY:
+        try:
+            user_state, created = MemberState.objects.get_or_create(user=user,
+                                                                    group=group)
 
-        if request.method == "POST":
-            form = VoteForm(request.POST, group_id=group.id)
-            if form.is_valid():
-                vote = form.save(commit=False)
-                vote.user = user
-                vote.group = group
+            if request.method == "POST":
+                form = VoteForm(request.POST, group_id=group.id)
+                if form.is_valid():
+                    vote = form.save(commit=False)
+                    vote.user = user
+                    vote.group = group
 
-                idea_vote1_id = (form.cleaned_data["idea_vote1"].id
-                                 if form.cleaned_data["idea_vote1"] else None)
-                idea_vote2_id = (form.cleaned_data["idea_vote2"].id
-                                 if form.cleaned_data["idea_vote2"] else None)
-                idea_vote3_id = (form.cleaned_data["idea_vote3"].id
-                                 if form.cleaned_data["idea_vote3"] else None)
+                    idea_vote1_id = (form.cleaned_data["idea_vote1"].id
+                                    if form.cleaned_data["idea_vote1"] else None)
+                    idea_vote2_id = (form.cleaned_data["idea_vote2"].id
+                                    if form.cleaned_data["idea_vote2"] else None)
+                    idea_vote3_id = (form.cleaned_data["idea_vote3"].id
+                                    if form.cleaned_data["idea_vote3"] else None)
 
-                idea_vote1 = Idea.objects.get(id=idea_vote1_id)
-                idea_vote2 = Idea.objects.get(id=idea_vote2_id)
-                idea_vote3 = Idea.objects.get(id=idea_vote3_id)
+                    idea_vote1 = Idea.objects.get(id=idea_vote1_id)
+                    idea_vote2 = Idea.objects.get(id=idea_vote2_id)
+                    idea_vote3 = Idea.objects.get(id=idea_vote3_id)
 
-                idea_vote1.votes += 1
-                idea_vote2.votes += 1
-                idea_vote3.votes += 1
+                    idea_vote1.votes += 1
+                    idea_vote2.votes += 1
+                    idea_vote3.votes += 1
 
-                idea_vote1.save()
-                idea_vote2.save()
-                idea_vote3.save()
+                    idea_vote1.save()
+                    idea_vote2.save()
+                    idea_vote3.save()
 
-                user_state.idea_vote1 = idea_vote1
-                user_state.idea_vote2 = idea_vote2
-                user_state.idea_vote3 = idea_vote3
-                user_state.save()
+                    user_state.idea_vote1 = idea_vote1
+                    user_state.idea_vote2 = idea_vote2
+                    user_state.idea_vote3 = idea_vote3
+                    user_state.save()
 
-                vote.save()
-                messages.success(request, "투표가 성공적으로 저장되었습니다.")
-                return redirect("group:group_detail", group_id=group_id)
-        else:
-            messages.error(request, "중복 선택은 불가능합니다.")
-            form = VoteForm(group_id=group_id)
+                    vote.save()
+                    messages.success(request, "투표가 성공적으로 저장되었습니다.")
+                    return redirect("group:group_detail", group_id=group_id)
+            else:
+                messages.error(request, "중복 선택은 불가능합니다.")
+                form = VoteForm(group_id=group_id)
 
-    except MemberState.DoesNotExist:
-        messages.error(request, "MemberState가 존재하지 않습니다.")
-        return redirect("group_detail", group_id=group_id)
+        except MemberState.DoesNotExist:
+            messages.error(request, "MemberState가 존재하지 않습니다.")
+            return redirect("group_detail", group_id=group_id)
 
-    voted_ideas = [
-        user_state.idea_vote1, user_state.idea_vote2, user_state.idea_vote3
-    ]
-    ideas_for_voting = (Idea.objects.filter(group=group).exclude(
-        author=user).exclude(
-            id__in=[idea.id for idea in voted_ideas if idea is not None]))
+        voted_ideas = [
+            user_state.idea_vote1, user_state.idea_vote2, user_state.idea_vote3
+        ]
+        ideas_for_voting = (Idea.objects.filter(group=group).exclude(
+            author=user).exclude(
+                id__in=[idea.id for idea in voted_ideas if idea is not None]))
 
-    return render(
-        request,
-        "group/group_vote_create.html",
-        {
-            "group": group,
-            "ideas_for_voting": ideas_for_voting,
-            "form": form
-        },
-    )
+        return render(
+            request,
+            "group/group_vote_create.html",
+            {
+                "group": group,
+                "ideas_for_voting": ideas_for_voting,
+                "form": form
+            },
+        )
 
 
 def result(request, group_id):  # 최종 결과 페이지
@@ -600,10 +622,11 @@ def result(request, group_id):  # 최종 결과 페이지
     idea_list = Idea.objects.filter(
         group=group).order_by("-score")[:group.team_number]
     members = MemberState.objects.filter(group=group)
+    state = redirect_by_auth(request.user, group_id)
 
-    ctx = {"idea_list": idea_list, "members": members, "group": group}
-
-    return render(request, "group/result.html", context=ctx)
+    if (state == State.WITH_HISTORY or state == State.ADMIN):
+        ctx = {"idea_list": idea_list, "members": members, "group": group}
+        return render(request, "group/result.html", context=ctx)
 
 
 def member_preresult(request, group_id):
@@ -612,21 +635,24 @@ def member_preresult(request, group_id):
         group=group).order_by("-score")[:group.team_number]
     user_state = MemberState.objects.filter(user=request.user,
                                             group=group).first()
-    ideas_votes = {}
-    if user_state:
-        ideas_votes["idea_vote1_id"] = (user_state.idea_vote1.id
-                                        if user_state.idea_vote1 else None)
-        ideas_votes["idea_vote2_id"] = (user_state.idea_vote2.id
-                                        if user_state.idea_vote2 else None)
-        ideas_votes["idea_vote3_id"] = (user_state.idea_vote3.id
-                                        if user_state.idea_vote3 else None)
-    ctx = {
-        "group": group,
-        "idea_list": idea_list,
-        "ideas_votes": ideas_votes,
-    }
+    state = redirect_by_auth(request.user, group_id)
 
-    return render(request, "preresult/preresult_member.html", ctx)
+    if (state == State.WITH_HISTORY or state == State.ADMIN):
+        ideas_votes = {}
+        if user_state:
+            ideas_votes["idea_vote1_id"] = (user_state.idea_vote1.id
+                                            if user_state.idea_vote1 else None)
+            ideas_votes["idea_vote2_id"] = (user_state.idea_vote2.id
+                                            if user_state.idea_vote2 else None)
+            ideas_votes["idea_vote3_id"] = (user_state.idea_vote3.id
+                                            if user_state.idea_vote3 else None)
+        ctx = {
+            "group": group,
+            "idea_list": idea_list,
+            "ideas_votes": ideas_votes,
+        }
+
+        return render(request, "preresult/preresult_member.html", ctx)
 
 
 @login_required
